refactor(plugins): apply ts naming guideline to src utility files (#741)
