+++ plugins/basemap-plugin/src/common/shared/viewportValidation.ts
