+++ plugins/folder-plugin/src/common/shared/folderValidation.ts
