+++ plugins/shape-plugin/src/common/types/index.ts
@@ -45,4 +45,4 @@ export {
   summarizeCheckboxState,
   getPreferredCountryCodeFormat,
   validateBuildConfig,
-} from '~/services/utils/utils';
+} from '~/services/utils/shapeBuildUtils';
