+++ plugins/shape-plugin/src/services/utils/shapeBuildUtils.ts
